Fix my_atof : better precision

+++ lib/include/my_conversions.h
@@ -14,7 +14,7 @@
 ** @return the number (0 if it not a number (my_atoi_err))
 **/
 int my_atoi(char const *str);
-float my_atof(char const *str);
+double my_atof(char const *str);
 double my_atodf(char const *str);
 
 /**
